review compose scripts and build
