+++ backend/Dockerfile.worker-mesh
@@ -28,9 +28,4 @@ COPY ./prestart.sh /app/
 COPY ./tests-start.sh /app/
 COPY ./app /app/app
 
-RUN mv /app/celery-reload.sh /celery-reload.sh && \
-    mv /app/celery.sh /celery.sh && \
-    chmod +x /celery-reload.sh && \
-    chmod +x /celery.sh
-
 ENTRYPOINT ["bash", "-c"]
