+++ backend/Dockerfile.worker-vector
@@ -55,9 +55,4 @@ COPY ./prestart.sh /app/
 COPY ./tests-start.sh /app/
 COPY ./app /app/app
 
-RUN mv /app/celery-reload.sh /celery-reload.sh && \
-    mv /app/celery.sh /celery.sh && \
-    chmod +x /celery-reload.sh && \
-    chmod +x /celery.sh
-
 ENTRYPOINT ["bash", "-c"]
