+++ backend/app/worker/pipelines/obj.py
@@ -2,4 +2,4 @@ from app.worker.main import celery
 
 
 def run(pipeline_extended):
-    return celery.send_task('create_mesh_3dtiles', kwargs={'pipeline_extended': pipeline_extended})
+    return celery.send_task('create_obj_mesh_3dtiles', kwargs={'pipeline_extended': pipeline_extended})
