--- backend/app/worker/tasks/cleanup/__init__.py
@@ -1,5 +1,0 @@
-QUEUE = 'cleanup'
-
-CELERY_INCLUDES = [
-    'app.worker.tasks.cleanup.task',
-]
