+++ backend/app/worker/tasks/common/__init__.py
@@ -1,0 +1,5 @@
+QUEUE = 'common'
+
+CELERY_INCLUDES = [
+    'app.worker.tasks.common.task',
+]
