+++ backend/app/worker/tasks/vector/task.py
@@ -1,3 +1,2 @@
 import app.worker.tasks.vector.pg2b3dm.task
 import app.worker.tasks.vector.i3dm.task
-import app.worker.tasks.cleanup.task
