--- backend/scripts/celery-reload.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-# using nodemon to be able to detect changes in the mounted volumes
-# the --legacy-watch is needed
-# other python solutions were not working properly
-nodemon -e py --legacy-watch --exec /celery.sh
