--- backend/scripts/celery.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-celery -A app.worker.main.celery worker --loglevel=info --concurrency=$CELERY_WORKER_CONCURRENCY
