+++ docker-compose.override.yml
@@ -1,7 +1,7 @@
 services:
 
   proxy:
-    image: traefik:3.0
+    image: traefik:v3.7.5
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
     ports:
