+++ docker-compose.worker.photogrammetry.yml
@@ -11,7 +11,7 @@ services:
     image: '${DOCKER_IMAGE_WORKER_PHOTOGRAMMETRY?Variable not set}:${TAG-latest}'
     platform: linux/amd64
     command:
-      - "celery -A app.worker.main.celery worker --queues=photogrammetry --include=app.worker.tasks.photogrammetry.task --loglevel=info --concurrency=${CELERY_WORKER_CONCURRENCY}"
+      - "celery -A app.worker.main.celery worker --queues=photogrammetry,common --include=app.worker.tasks.photogrammetry.task,app.worker.tasks.common.task --loglevel=info --concurrency=${CELERY_WORKER_CONCURRENCY}"
     env_file:
       - .env
     networks:
